requirements.txt
IMDbPY==2021.4.18
pandas==1.4.2
Pillow==9.1.1
requests==2.27.1
streamlit==1.9.0
